+++ src/test/resources/ru/vyarus/gradle/plugin/quality/gen/Partner.java
@@ -1,0 +1,25 @@
+package com.something;
+
+/**
+ * Dummy comment.
+ */
+public class Partner {
+    private Boolean dummyAccount;
+
+    public Partner() {
+        // intentionally empty
+    }
+
+    public Partner(final Boolean dummyAccount) {
+        this.dummyAccount = dummyAccount;
+    }
+
+    public Boolean getDummyAccount() {
+        return dummyAccount;
+    }
+
+    public void setDummyAccount(final Boolean dummyAccount) {
+        this.dummyAccount = dummyAccount;
+    }
+}
+