+++ src/test/resources/ru/vyarus/gradle/plugin/quality/gen/PartnerAccount.java
@@ -1,0 +1,24 @@
+package com.something;
+
+/**
+ * Dummy comment.
+ */
+public class PartnerAccount {
+    private PartnerAccountData partnerAccountData;
+
+    public PartnerAccount() {
+        // intentionally empty
+    }
+
+    public PartnerAccount(final PartnerAccountData partnerAccountData) {
+        this.partnerAccountData = partnerAccountData;
+    }
+
+    public PartnerAccountData getPartnerAccountData() {
+        return partnerAccountData;
+    }
+
+    public void setPartnerAccountData(final PartnerAccountData partnerAccountData) {
+        this.partnerAccountData = partnerAccountData;
+    }
+}