+++ src/test/resources/ru/vyarus/gradle/plugin/quality/java/sample/ValidSample.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class ValidSample {
+
+    public static void main(String[] args) {
+        System.out.println("Hello world");
+    }
+}